Improved code structure of button and added css module (#11)
* Improved code structure of button and added css module

* Fixed DefaultButtons import statement in App.tsx

---------
