+++ frontend/src/App.tsx
@@ -1,7 +1,7 @@
 import './App.css'
 import {NavigationMenu} from "./Components/NavigationMenu.tsx";
 import {SearchField} from "./Components/SearchField.tsx";
-import DefaultButton from "./Components/Button.tsx";
+import DefaultButton from "./Components/DefaultButton.tsx";
 import InputField from "./Components/InputField.tsx";
 import {UpperMenu} from "./Components/UpperMenu.tsx";
