+++ frontend/src/Styling/DefaultButton.module.css
@@ -1,0 +1,45 @@
+.defaultButton {
+    font-family: 'Verdan', sans-serif;
+    font-weight: bold;
+    text-transform: uppercase;
+    letter-spacing: 0.5px;
+    border: none;
+    border-radius: 4px;
+    cursor: pointer;
+    transition: all 0.3s ease;
+    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+}
+
+.defaultButton:hover {
+    opacity: 0.9;
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.15);
+}
+
+.defaultButton:active {
+    transform: translateY(1px);
+    box-shadow: 0 1px 2px rgba(0, 0, 0, 0.1);
+}
+
+.defaultButton:disabled {
+    opacity: 0.5;
+    cursor: not-allowed;
+}
+
+.small {
+    font-size: 12px;
+    padding: 6px 12px;
+}
+
+.medium {
+    font-size: 14px;
+    padding: 8px 16px;
+}
+
+.large {
+    font-size: 16px;
+    padding: 10px 20px;
+}
+
+.fullWidth {
+    width: 100%;
+}
